+++ src/main/java/com/ec/final_project/Services/taikhoan_thanhtoanService.java
@@ -1,6 +1,6 @@
 package com.ec.final_project.Services;
 
 public interface taikhoan_thanhtoanService {
-    void addtkTT(int sotien,String sotaikhoan,int acc_id);
-    void updatetkTT(int sotien,int acc_id);
+    void addtkTT(double sotien,String sotaikhoan,int acc_id);
+    void updatetkTT(double sotien,int acc_id);
 }
